sketch CPA core algo
